Remove a race condition between a child process reaping, t.i.process.Process notifying of the process exit and t.r.procmon.ProcessRunner trying to send a signal to the child process. The problem is that Process only notifies of the child's exit after the childs stdout and stderr have seen EOF, but the pid is invalid from the moment of the reaping. Refactor procmon to not see any pids ever, but use Process.signalProcess() instead. Add a new exception to t.i.process that tells that the process you are trying to signal has already been reaped.
git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@5735 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb

+++ twisted/internet/process.py
@@ -176,6 +176,9 @@ class ProcessError(abstract.FileDescriptor):
         os.close(self.proc.stderr)
         self.proc.errConnectionLost()
 
+class ProcessExitedAlready(Exception):
+    """The process has already excited, and the operation requested can no longer be performed."""
+    pass
 
 class Process(abstract.FileDescriptor, styles.Ephemeral):
     """An operating-system Process.
@@ -330,6 +333,8 @@ class Process(abstract.FileDescriptor, styles.Ephemeral):
     def signalProcess(self, signalID):
         if signalID in ('HUP', 'STOP', 'INT', 'KILL'):
             signalID = getattr(signal, 'SIG'+signalID)
+        if self.pid is None:
+            raise ProcessExitedAlready
         os.kill(self.pid, signalID)
 
     def doError(self):
@@ -392,6 +397,7 @@ class Process(abstract.FileDescriptor, styles.Ephemeral):
     def processEnded(self, status):
         self.status = status
         self.lostProcess = 1
+        self.pid = None
         self.closeStdin()
         self.maybeCallProcessEnded()
